feat: eslint comma 오류 해결 및 api 폴더 구조 변경
